+++ footmark/__init__.py
@@ -5,7 +5,7 @@ import os
 
 from footmark.pyami.config import Config, FootmarkLoggingConfig, DefaultLoggingConfig
 
-__version__ = '1.3.0'
+__version__ = '1.5.0'
 Version = __version__  # for backware compatibility